agent_tools/jira_mcp_server/requirements.txt
mcp==1.13.1
jira>=3.10.0
python-dotenv>=1.0.0
pydantic>=1.10.0
